@@ -1,26 +1,25 @@
 name: youtube_player_iframe
 description: Flutter port of the official YouTube iFrame player API. Supports web & mobile platforms.
-version: 4.0.4
+version: 5.1.1
 repository: https://github.com/sarbagyastha/youtube_player_flutter
 homepage: https://github.com/sarbagyastha/youtube_player_flutter/tree/master/packages/youtube_player_iframe
 
 environment:
   sdk: '>=3.0.0 <4.0.0'
-  flutter: '>=3.0.0'
+  flutter: '>=3.18.0-0'
 
 dependencies:
   flutter:
     sdk: flutter
   meta: '>=1.7.0 <2.0.0'
-  webview_flutter: ^4.2.2
-  webview_flutter_android: ^3.9.2
-  webview_flutter_wkwebview: ^3.7.2
-  url_launcher: ^6.1.12
-  youtube_player_iframe_web: ^2.0.2
+  webview_flutter: ^4.7.0
+  webview_flutter_android: ^3.15.0
+  webview_flutter_wkwebview: ^3.12.0
+  url_launcher: ^6.2.4
+  youtube_player_iframe_web: ^3.0.1
 
 dev_dependencies:
-  flutter_test:
-    sdk: flutter
+  flutter_lints: ^3.0.1
 
 flutter:
   assets:
